+++ tests/run.lisp
@@ -13,6 +13,7 @@
 (load (project-file "src/application/ports.lisp"))
 (load (project-file "src/application/create-run.lisp"))
 (load (project-file "src/application/arm-run.lisp"))
+(load (project-file "src/application/start-run.lisp"))
 
 (in-package :hngh.tests)
 
@@ -35,6 +36,7 @@
 (load (cl-user::project-file "tests/support/fakes.lisp"))
 (load (cl-user::project-file "tests/application/test-create-run.lisp"))
 (load (cl-user::project-file "tests/application/test-arm-run.lisp"))
+(load (cl-user::project-file "tests/application/test-start-run.lisp"))
 
 (defun read-fixture (relative)
   (with-open-file (stream (cl-user::project-file relative))